app.jsx and navbar fixed
