+++ src/components/Navbar/Navbar.module.css
@@ -1,71 +1,90 @@
-::-webkit-scrollbar {
-  width: 10px;
-}
-
-/* Track */
-::-webkit-scrollbar-track {
-  background: black;
-}
-
-/* Handle */
-::-webkit-scrollbar-thumb {
-  background: white;
-}
-
-/* Handle on hover */
-::-webkit-scrollbar-thumb:hover {
-  background: orange;
-}
-
 .navbar {
-  position: sticky;
-  top: 0;
-  z-index: 1;
   display: flex;
   justify-content: space-between;
   align-items: center;
-  padding: 2px 20px;
+  padding: 20px 60px;
+}
+.navlist {
+  display: flex;
+  justify-content: center;
+  align-items: center;
+  gap: 0 30px;
+  padding: 0;
+}
+.list_item {
+  list-style-type: none;
+  font-size: 18px;
+  font-weight: 600;
+  cursor: pointer;
+  position: relative;
+  display: inline-block;
+}
+.list_item::after {
+  content: "";
+  position: absolute;
+  width: 100%;
+  height: 2px;
+  bottom: -5px;
+  left: 0;
+  background-color: blue;
+  visibility: hidden;
+  transform: scaleX(0);
+  transition: all 0.3s ease-in-out;
 }
 
-.navbar-light {
-  background-color: rgba(256, 256, 256, 1);
-  box-shadow: 0px 7px 18px 0px #0000002e;
+.list_item:hover::after {
+  visibility: visible;
+  transform: scaleX(1);
 }
-
-.navbar-dark {
-  background-color: rgb(20, 20, 20);
-  box-shadow: 0px 7px 18px 0px #ffffff2e;
+.list_item_link {
+  text-decoration: none;
+  color: #000;
 }
-
-.logo {
-  width: 100px;
+.navbar.navbar-dark .navlist .list_item .list_item_link {
+  color: #fff;
 }
 .github {
-  width: 52px;
-  position: relative;
-  top: 3px;
-  height: 52px;
+  background-color: #4285f4;
+  width: 200px;
+  height: 60px;
+  border-radius: 12px;
+  padding: 10px;
+  display: flex;
+  justify-content: space-around;
+  align-items: center;
 }
-.github,
-.spider {
-  width: 50px;
-  filter: grayscale(1);
-  transition: all 0.3s ease;
+.image_container {
+  width: 40px;
+  height: 40px;
+  border-radius: 50%;
+  overflow: hidden;
+  border: 1px solid #fff;
+  box-shadow: 0px 4px 1px 0px #00000040;
 }
-.github:hover,
-.spider:hover {
-  filter: grayscale(0);
-  transform: rotate(-10deg) scale(1.2);
+
+.image {
+  width: 100%;
+  height: 100%;
+  object-fit: cover;
 }
-.spider_img {
-  width: 80px;
+.username {
+  font-size: 18px;
+  font-weight: 600;
+  color: #000;
+  text-shadow: 0px 4px 4px 0px #00000040;
+  filter: drop-shadow(0px 4px 1px rgba(255, 255, 255, 0.25));
+}
+.navbar.navbar-dark .username {
+  color: #fff;
+  text-shadow: 0px 4px 4px 0px #ffffff40;
+  filter: drop-shadow(0px 4px 1px rgba(0, 0, 0, 0.25));
 }
 .mode_toggle {
   position: relative;
   display: block;
   border: none;
-  width: 50px;
-  height: 50px;
+  width: 40px;
+  height: 40px;
   padding: 10px;
   border-radius: 300px;
   margin-left: auto;
@@ -83,7 +102,6 @@
   color: var(--dark-text-color);
   box-shadow: 3px 3px 5px 0px #828282;
 }
-
 .light_mode img {
   filter: invert(100%);
 }
