+++ src/index.css
@@ -4,7 +4,7 @@
   --light-text-color: #1a1a1a;
 
   /* dark theme */
-  --dark-bg-color: #2e2e2e;
+  --dark-bg-color: #151515;
   --dark-text-color: #fff;
 }
